+++ src/mctpd.c
@@ -1601,7 +1601,7 @@ static int peer_set_mtu(struct ctx *ctx, struct peer *peer, uint32_t mtu)
 		return -EPROTO;
 	}
 
-	rc = mctp_nl_route_del(ctx->nl, peer->eid, peer->phys.ifindex);
+	rc = mctp_nl_route_del(ctx->nl, peer->eid, peer->phys.ifindex, NULL);
 	if (rc < 0 && rc != -ENOENT) {
 		warnx("%s, Failed removing existing route for eid %d %s",
 		      __func__, peer->phys.ifindex,
@@ -1609,7 +1609,8 @@ static int peer_set_mtu(struct ctx *ctx, struct peer *peer, uint32_t mtu)
 		// Continue regardless, route_add will likely fail with EEXIST
 	}
 
-	rc = mctp_nl_route_add(ctx->nl, peer->eid, peer->phys.ifindex, mtu);
+	rc = mctp_nl_route_add(ctx->nl, peer->eid, peer->phys.ifindex, NULL,
+			       mtu);
 	if (rc >= 0) {
 		peer->mtu = mtu;
 	}
@@ -2324,10 +2325,10 @@ static int peer_route_update(struct peer *peer, uint16_t type)
 
 	if (type == RTM_NEWROUTE) {
 		return mctp_nl_route_add(peer->ctx->nl, peer->eid,
-					 peer->phys.ifindex, peer->mtu);
+					 peer->phys.ifindex, NULL, peer->mtu);
 	} else if (type == RTM_DELROUTE) {
 		return mctp_nl_route_del(peer->ctx->nl, peer->eid,
-					 peer->phys.ifindex);
+					 peer->phys.ifindex, NULL);
 	}
 
 	bug_warn("%s: bad type %d", __func__, type);